feat: publicly expose error types
