+++ internal/http/client.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/newrelic/newrelic-client-go/internal/version"
 	"github.com/newrelic/newrelic-client-go/pkg/config"
+	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
 const (
@@ -228,7 +229,7 @@ func (c *NewRelicClient) do(
 	defer resp.Body.Close()
 
 	if resp.StatusCode == http.StatusNotFound {
-		return nil, &ErrorNotFound{}
+		return nil, &errors.ErrorNotFound{}
 	}
 
 	body, readErr := ioutil.ReadAll(resp.Body)
@@ -240,9 +241,9 @@ func (c *NewRelicClient) do(
 		errorValue := c.errorValue
 		_ = json.Unmarshal(body, &errorValue)
 
-		return nil, &ErrorUnexpectedStatusCode{
-			statusCode: resp.StatusCode,
-			err:        c.errorValue.Error()}
+		return nil, &errors.ErrorUnexpectedStatusCode{
+			StatusCode: resp.StatusCode,
+			Err:        c.errorValue.Error()}
 	}
 
 	if value == nil {
