+++ internal/http/errors.go
@@ -30,29 +30,3 @@ func (e *DefaultErrorResponse) Error() string {
 
 	return m
 }
-
-// ErrorNotFound is returned when a 404 response is returned
-// from New Relic's APIs.
-type ErrorNotFound struct{}
-
-func (e *ErrorNotFound) Error() string {
-	return fmt.Sprintf("404 not found")
-}
-
-// ErrorUnexpectedStatusCode is returned when an unexpected
-// status code is returned from New Relic's APIs.
-type ErrorUnexpectedStatusCode struct {
-	err        string
-	statusCode int
-}
-
-func (e *ErrorUnexpectedStatusCode) Error() string {
-
-	msg := fmt.Sprintf("%d response returned", e.statusCode)
-
-	if e.err != "" {
-		msg += fmt.Sprintf(": %s", e.err)
-	}
-
-	return msg
-}
